Move validation stylesheet functions to picard/ui/colors.py
Move stylesheet_validation_success() and stylesheet_validation_error()
from picard/ui/options/__init__.py to picard/ui/colors.py where they
logically belong alongside other color definitions. This avoids the
need to import from the options module just for color support in
non-options dialogs.
